Refactoring Complete
